Isolate FA3 batch invariance regression test
Keep the H100 regression independent of the broader copied vLLM test module.\n\nAssisted-by: OpenAI Codex
